Resolved linux build issue.
Using uint64_t instead of size_t to guarantee largest int size regardless of system. Including the c++ equivalents to stdlib.h, cstdlib and cstring.

+++ cluster_dynamics.hpp
@@ -148,8 +148,8 @@ extern int simulation_time;
 extern int delta_time;
 
 // result arrays
-extern size_t* interstitials;
-extern size_t* vacancies;
+extern uint64_t* interstitials;
+extern uint64_t* vacancies;
 
 extern NuclearReactor* reactor;
 extern Material* material;
